code: move more things
